\encoding{UTF-8}
\name{dplR-defunct}
\alias{dplR-defunct}
\alias{plotRings}
\alias{ffcsaps}
\title{
  Defunct functions in dplR
}
\description{
  These are functions no longer included in dplR
}
\details{
  \tabular{ll}{
    Function: \tab plotRings\cr
    Function: \tab ffcsaps, use \code{\link{caps}} instead\cr
  }
}
\keyword{ help }
