% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPLR.R
\name{dPLR}
\alias{dPLR}
\title{dPLR: Distributed Proportional Likelihood Ratio Model}
\usage{
dPLR(mydata, id.local=1, init_est = 'meta',  output.dPLR1 = F, control=list(maxit=100), verbose = F)
}
\arguments{
\item{mydata}{Data frame of the data from all sites, with the first three columns as site id, outcome,
offset (e.g. observation time) and the rest columns as covariates.}

\item{id.local}{local site id}

\item{init_est}{Character or numeric vector, default 'meta' which uses PLR meta estimate as initial beta_bar.}

\item{output.dPLR1}{Logical, if TRUE the surrogate likelihood is the 1st order approximation and PLR1 is also fitted. Default FALSE.}

\item{control}{control options passed to \code{optim()}, e.g. list(maxit=100).}

\item{verbose}{Logical - show running details?}
}
\value{
List with components:

\item{beta_bar}{initial coef estimate, by local site estimate or meta estimate}

\item{res_plr}{PLR fitted at all sites}

\item{b_meta}{meta PLR estimate}

\item{b_local}{local site PLR estimate}

\item{beta_tilde}{coef estimate by optimizing the distPLR surrogate likelihood function (2nd order approximation)}

\item{var_tilde}{var estimate of beta_tilde calculated from hessian}

\item{beta_tilde1}{coef estimate by optimizing the distPLR surrogate likelihood function (1st order approximation)}

\item{var_tilde1}{var estimate of beta_tilde1 calculated from hessian}

\item{sol}{an \code{optim()} output, optimizing distPLR surrogate likelihood function (2nd order approximation)}

\item{sol1}{an \code{optim()} output, optimizing distPLR surrogate likelihood function (1st order approximation)}

\item{grad_LN_L1_beta_bar}{sum gradients from collaborative sites, used in constructing the distPLR surrogate likelihood}

\item{hess_LN_L1_beta_bar}{sum hessians from collaborative sites, used in constructing the distPLR surrogate likelihood}
}
\description{
Fit a privacy-preserving and communication-efficient dPLR model to integrate data across Clinical Sites. See Luo, et al. 2023 for details.
             This function shows the estimation accuracy of the method: dPLR is closer to the stratified PLR (pooled analysis) compared to the meta-estimator.
}
\details{
The surrogate pairwise log-likelihood is the logL of local site plus
         the 1st and 2nd order approximation of the logL of remote sites
         Ltilde(beta) = L1(beta) + sum(beta * grad(LN(beta_bar) - L1(beta_bar)))+ t(beta) %*% hessian(LN(beta_bar) - L1(beta_bar)) %*% beta / 2.
         It's optimized by \code{optim()}. The log-L, gradients and hessian are written in Rcpp.
}
\examples{
data(sim_data_dPLR)
## distPLR
fit_dist <- dPLR(mydata=sim_data_dPLR, id.local = 1, init_est = 'meta')
b_plr_dist <- c(fit_dist$beta_tilde)
b_plr_meta <- fit_dist$b_meta

## pooling data together: stratified PLR (as gold-standard)
fit_strat = stratPLR(mydata=sim_data_dPLR)
b_plr_strat <- fit_strat$par

## dPLR is closer to gold-standard compared to the meta-estimator
round(cbind(b_plr_strat, b_plr_meta, b_plr_dist), 3)
}
\references{
Chongliang Luo, et al. "Distributed Proportional Likelihood Ratio Model with Application to Data Integration across Clinical Sites." AoAS (2023).
            Xiaodong Luo and Wei Yann Tsai. "A proportional likelihood ratio model." Biometrika (2012).
             Michael I. Jordan, Jason D. Lee, and Yun Yang. "Communication-efficient distributed statistical inference." JASA (2018).
}
\author{
Chongliang Luo
}
