\encoding{UTF-8}
\name{i.detrend}
\alias{i.detrend}
\title{ Interactively Detrend Multiple Ring-Width Series }
\description{
  Interactively detrend multiple tree-ring series by one of two methods, a
  smoothing spline or a statistical model.  This is a wrapper for
  \code{\link{detrend.series}}.
}
\usage{
i.detrend(rwl, y.name = names(rwl), nyrs = NULL, f = 0.5,
          pos.slope = FALSE)
}
\arguments{
  \item{rwl}{ a \code{data.frame} with series as columns and years as
    rows such as that produced by \code{\link{read.rwl}} or
    \code{\link{ca533}} }
  \item{y.name}{ a \code{character} vector of length
    \code{ncol(\var{rwl})} that gives the \acronym{ID} of each
    series.  Defaults to the column names of \code{\var{rwl}}. }
  \item{nyrs}{ a number giving the rigidity of the smoothing spline,
    defaults to 0.67 of series length if \code{\var{nyrs}} is
    \code{NULL}. }
  \item{f}{ a number between 0 and 1 giving the frequency response or
    wavelength cutoff.  Defaults to 0.5. }
  \item{pos.slope}{ a \code{logical} flag.  Will allow for a positive
    slope to be used in method \code{"ModNegExp"}.  If \code{FALSE} the
    line will be horizontal. }
}
\details{
  This function allows a user to choose detrending curves based on plots
  that are produced by \code{\link{detrend.series}} for which it is
  essentially a wrapper.  The user enters their choice of detrended
  method via keyboard at a prompt for each ring width series in
  \code{\var{rwl}}.  See \code{\link{detrend.series}} for examples and
  details on the detrending methods.

  A series with no values at all is dropped, with a message naming it,
  before any series is shown, and its name is dropped from
  \code{\var{y.name}} too.
}
\value{
  An object of class \code{c("rwi", "data.frame")} (see
  \code{\link{as.rwi}}) containing each detrended series according to
  the method used as columns and rownames set to
  \code{colnames(\var{y})}.  These are typically years.  The method
  chosen for each series is recorded, named by series, in
  \code{attr(x, "dplR.detrend")$method}.  Plots are also
  produced as the user chooses the detrending methods through keyboard
  input.
}
\note{
  This function is kept for old scripts.  For interactive detrending use
  the iDetrend app (\url{https://github.com/OpenDendro/iDetrend}), part
  of openDendro.  There the curve for each series can be adjusted and
  compared with others on screen, series that need attention are pointed
  out, and the app writes \R code that reproduces the indices.  Here the
  choice is made once per series from a static plot, the parameters of
  the curves cannot be changed series by series, and only the name of
  the method chosen is recorded.  The function says so, once in a
  session, when it is called.
}
\author{ Andy Bunn }
\seealso{ \code{\link{detrend.series}} }
\keyword{ iplot }
