% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPLR.R
\name{stratPLR}
\alias{stratPLR}
\title{stratPLR: stratified Proportional Likelihood Ratio Model}
\usage{
stratPLR(mydata)
}
\arguments{
\item{mydata}{Data frame of the data from all sites, with the first three columns as site id, outcome,
offset (e.g. observation time) and the rest columns as covariates.}
}
\value{
{an \code{optim()} output, optimizing the stratified PLR likelihood function.}
}
\description{
Fit a stratified PLR to integrate data across clinical sites.
The stratPLR likelihood does not involve the nuisance baseline distribution
functions, and hence allow the baseline distributions to be different from
sites. See Luo, et al. 2023 for details.
}
\references{
Chongliang Luo, et al. "Distributed Proportional Likelihood Ratio Model with Application to Data Integration across Clinical Sites." AoAS (2023).
            Xiaodong Luo and Wei Yann Tsai. "A proportional likelihood ratio model." Biometrika (2012).
}
\author{
Chongliang Luo
}
