\encoding{UTF-8}
\name{zof.anc}
\docType{data}
\alias{zof.anc}
\title{ Ancillary Data Corresponding to \code{\link{zof.rwl}} }
\description{
  This data set gives the pith offsets, distance to pith, and diameter that match the ring widths for \code{\link{zof.rwl}} \enc{–}{-} a data set of XXXX (\emph{XXXX}) from the XXXX in XXXX.
}
\usage{data(zof.anc)}
\format{
  A \code{data.frame} containing four columns. Column one gives the series ID. Column two (\code{PO}) gives the number of rings estimated to be missing to the pith. Column three (\code{d2pith}) gives the estimated distance to the pith (mm). Column four (\code{diam}) gives the diameter at breast height (DBH) in cm.
}
\source{
Stefan Klesse
}
\keyword{datasets}
