\encoding{UTF-8}
\name{zof.rwl}
\docType{data}
\alias{zof.rwl}
\title{ XXXX }
\description{
  This data set includes ring-width measurements for XXXXX (\emph{XXX XXXX}) increment cores collected at the XXXX in XXXXX.  There are 61 series.
}
\usage{data(zof.rwl)}
\format{
  A \code{data.frame} containing 61 ring-width series in columns and 156
  years in rows.
}
\source{
  Stefan Klesse
}
\keyword{datasets}
