\name{anos1}
\docType{data}
\alias{anos1}
\title{Rothenburg Tree Ring Widths}
\description{
  This dataset gives the raw ring widths for Norway spruce \emph{Picea
  abies} at Rothenburg ob der Tauber, Bavaria, Germany.  There are 20
  series from 10 trees.  Dataset was created using
  \code{\link{read.rwl}} and saved to an .rda file using
  \code{\link[base]{save}}.
}
\usage{data(anos1)}
\format{A \code{data.frame} containing 20 tree-ring series from 10 trees
  in columns and 98 years in rows. The correct stc mask for use with
  \code{\link{read.ids}} is \code{c(5, 2, 1)}.}
\references{
  Zang, C. (2010) Growth reaction of temperate forest tree species to
  summer drought -- a multispecies tree-ring network
  approach. PhD-Thesis, Technische Universität München.
}
\keyword{datasets}
