\name{cana157}
\docType{data}
\alias{cana157}
\title{Twisted Tree Heartrot Hill Standard Chronology}
\description{
  This dataset gives the standard chronology for white spruce
  \emph{Picea glauca} at Twisted Tree Heartrot Hill in Yukon,
  Canada.  Dataset was created using \code{\link{read.crn}} and saved to
  an .rda file using \code{\link{save}}.
}
\usage{
cana157
}
\format{
  A \code{data.frame} containing the standard chronology in column one
  and the sample depth in column two.  There are 463 years (1530--1992)
  in the rows.
}
\source{International tree-ring data bank, Accessed on 27-August-2007 at
  \url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/treering/chronologies/northamerica/canada/cana157.crn}
}
\references{
  Jacoby, G., D'Arrigo, R. and Buckley, B. (1992) Twisted Tree Heartrot
  Hill Data Set.  \acronym{IGBP} \acronym{PAGES}/World Data Center for
  Paleoclimatology Data Contribution Series 1992-CANA157.CRN.
  \acronym{NOAA}/\acronym{NCDC} Paleoclimatology Program, Boulder,
  Colorado, \acronym{USA}.
}
\keyword{datasets}
