\name{co021}
\docType{data}
\alias{co021}
\title{Schulman Old Tree No. 1, Mesa Verde}
\description{
  This dataset gives the raw ring widths for Douglas fir
  \emph{Pseudotsuga menziesii} at Mesa Verde in Colorado, USA. There are 35
  series. Dataset was created using \code{\link{read.rwl}} and saved to an .rda
  file using \code{\link[base]{save}}.
}
\usage{co021}
\format{A \code{data.frame} containing 35 tree-ring series in columns and 788 years
  in rows. }
\source{International tree-ring data bank, Accessed on 27-August-2007 at
  \url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/treering/chronologies/northamerica/usa/co021.crn}
}
\references{
  Schulman, E. (1963) Schulman Old Tree No. 1 Data Set. IGBP PAGES/World Data
  Center for Paleoclimatology Data Contribution Series 1983-CO021.RWL.
  NOAA/NCDC Paleoclimatology Program, Boulder, Colorado, USA.
}
\keyword{datasets}
