\name{co021}
\docType{data}
\alias{co021}
\title{Schulman Old Tree No. 1, Mesa Verde}
\description{
  This dataset gives the raw ring widths for Douglas fir
  \emph{Pseudotsuga menziesii} at Mesa Verde in Colorado, \acronym{USA}.
  There are 35 series.  Dataset was created using \code{\link{read.rwl}}
  and saved to an .rda file using \code{\link{save}}.
}
\usage{
co021
}
\format{
  A \code{data.frame} containing 35 tree-ring series in columns and 788
  years in rows.
}
\source{
  International tree-ring data bank, Accessed on 12-September-2011 at
  \url{ftp://ftp.ncdc.noaa.gov/pub/data/paleo/treering/measurements/northamerica/usa/co021.rwl}
}
\references{
  Schulman, E. (1963) Schulman Old Tree No. 1 Data Set.  \acronym{IGBP}
  \acronym{PAGES}/World Data Center for Paleoclimatology Data
  Contribution Series 1983-CO021.RWL.  \acronym{NOAA}/\acronym{NCDC}
  Paleoclimatology Program, Boulder, Colorado, \acronym{USA}.
}
\keyword{datasets}
