\name{gp.d2pith}
\docType{data}
\alias{gp.d2pith}
\title{ Ponderosa pine distance to pith corresponding to gp.rwl }
\description{
This dataset gives the distance to pith for each series (in mm) that matches the ring widths for \code{gp.rwl} - a dataset of ponderosa pine (\emph{Pinus ponderosa}) from the Gus Pearson Natural Area (GPNA) in northern Arizona, USA. Data are further described by Biondi and Qeadan (2008) and references therein.
}
\usage{gp.d2pith}
\format{A \code{data.frame} containing series ids in column 1 (\code{series}) and the distance to the center of the tree in (\code{d2pith}). This can be used together with the ring widths to calculate the area of each ring.
}
\source{DendroLab, University of Nevada Reno, USA.
  \url{http://dendrolab.org/}
}
\references{ Biondi, F. and Qeadan, F. (2008) A theory-driven approach to tree-ring
  standardization: Defining the biological trend from expected basal area
  increment. Tree-Ring Research 64(2): 81-96. }
\keyword{datasets}
