\name{gp.po}
\docType{data}
\alias{gp.po}
\title{ Ponderosa Pine Pith Offsets Corresponding to gp.rwl }
\description{
This dataset gives the pith offsets that match the ring widths for \code{gp.rwl} - a dataset of ponderosa pine (\emph{Pinus ponderosa}) from the Gus Pearson Natural Area (GPNA) in northern Arizona, USA. Data are further described by Biondi and Qeadan (2008) and references therein.
}
\usage{gp.po}
\format{A \code{data.frame} containing series ids in column 1 (\code{series}) and the number of years between the beginning of that series in \code{gp.rwl} and the pith of the tree (\code{pith.offset}). This can be used together with the ring widths to calculate the cambial age of each ring.
}
\source{DendroLab, University of Nevada Reno, USA.
  \url{http://dendrolab.org/}
}
\references{ Biondi, F. and Qeadan, F. (2008) A theory-driven approach to tree-ring
  standardization: Defining the biological trend from expected basal area
  increment. Tree-Ring Research 64(2): 81-96. }
\keyword{datasets}
