\encoding{UTF-8}
\name{gp.rwl}
\docType{data}
\alias{gp.rwl}
\title{ Ponderosa Pine Ring Widths from Gus Pearson Natural Area }
\description{
  This data set includes ring-width measurements for ponderosa pine
  (\emph{Pinus ponderosa}) increment cores collected at the Gus Pearson
  Natural Area (\acronym{GPNA}) in northern Arizona,
  \acronym{USA}.  There are 58 series from 29 trees (2 cores per
  tree).  Data are further described by Biondi and Qeadan (2008) and
  references therein.
}
\usage{data(gp.rwl)}
\format{
  A \code{data.frame} containing 58 ring-width series in columns and 421
  years in rows.
}
\source{
  DendroLab, University of Nevada Reno, \acronym{USA}.
  \url{http://dendrolab.org/}
}
\references{
  Biondi, F. and Qeadan, F. (2008) A theory-driven approach to tree-ring
  standardization: Defining the biological trend from expected basal area
  increment.  \emph{Tree-Ring Research}, \bold{64}(2), 81\enc{–}{--}96.
}
\keyword{datasets}
