\name{gp.rwl}
\docType{data}
\alias{gp.rwl}
\title{ Ponderosa Pine Ring Widths from Gus Pearson Natural Area }
\description{
  This dataset includes ring-width measurements for ponderosa pine
  (\emph{Pinus ponderosa}) increment cores collected at the Gus Pearson
  Natural Area (\acronym{GPNA}) in northern Arizona,
  \acronym{USA}. There are 58 series from 29 trees (2 cores per
  tree). Data are further described by Biondi and Qeadan (2008) and
  references therein.
}
\usage{gp.rwl}
\format{
  A \code{data.frame} containing 58 ring-width series in columns and 421
  years in rows.
}
\source{
  DendroLab, University of Nevada Reno, \acronym{USA}.
  \url{http://dendrolab.org/}
}
\references{
  Biondi, F. and Qeadan, F. (2008) A theory-driven approach to tree-ring
  standardization: Defining the biological trend from expected basal
  area increment.  \emph{Tree-Ring Research}, 64(2):81--96.
}
\keyword{datasets}
