\encoding{UTF-8}
\name{nm046}
\docType{data}
\alias{nm046}
\title{Los Alamos Tree Ring Widths}
\description{
  This data set gives the raw ring widths for Douglas fir
  \emph{Pseudotsuga menziesii} near Los Alamos New Mexico,
  \acronym{USA}.  There are 8 series.  Data set was created using
  \code{\link{read.rwl}} and saved to an .rda file using
  \code{\link{save}}.
}
\usage{data(nm046)}
\format{
  A \code{data.frame} containing 8 tree-ring series in columns and 289
  years in rows.
}
\source{
  International tree-ring data bank, Accessed on 20-April-2021 at
  \url{https://www.ncei.noaa.gov/pub/data/paleo/treering/measurements/northamerica/usa/nm046.rwl}
}
\references{
  O'Brien, D (2002) Los Alamos Data
  Set.  \acronym{IGBP} \acronym{PAGES}/World Data Center for
  Paleoclimatology Data Contribution Series 2002-NM046.RWL.
  \acronym{NOAA}/\acronym{NCDC} Paleoclimatology Program, Boulder,
  Colorado, \acronym{USA}.
}
\keyword{datasets}
