\encoding{UTF-8}
\name{wa082}
\docType{data}
\alias{wa082}
\title{Hurricane Ridge, Pacific silver fir}
\description{
  This data set gives the raw ring widths for Pacific silver fir
  \emph{Abies amabilis} at Hurricane Ridge in Washington, \acronym{USA}.
  There are 23 series.  Data set was created using \code{\link{read.rwl}}
  and saved to an .rda file using \code{\link{save}}.
}
\usage{data(wa082)}
\format{
  A \code{data.frame} containing 23 tree-ring series in columns and 286
  years in rows.
}
\source{
  International tree-ring data bank, Accessed on 20-April-2021 at
  \url{https://www.ncei.noaa.gov/pub/data/paleo/treering/measurements/northamerica/usa/wa082.rwl}
}
\references{
  Schweingruber, F. (1983) Hurricane Ridge Data Set.  \acronym{IGBP}
  \acronym{PAGES}/World Data Center for Paleoclimatology Data
  Contribution Series 1983-wa082.RWL.  \acronym{NOAA}/\acronym{NCDC}
  Paleoclimatology Program, Boulder, Colorado, \acronym{USA}.
}
\keyword{datasets}
