\encoding{UTF-8}
\name{zof.anc}
\docType{data}
\alias{zof.anc}
\title{ Ancillary Data Corresponding to \code{\link{zof.rwl}} }
\description{
  This data set gives the pith offsets, distance to pith, and diameter that match the ring widths for \code{\link{zof.rwl}} \enc{–}{-} a data set of European Beech (\emph{Fagus sylvatica}) increment cores collected at the Zofingen in Switzerland.
}
\usage{data(zof.anc)}
\format{
  A \code{data.frame} containing four columns. Column one gives the series ID. Column two (\code{PO}) gives the number of rings estimated to be missing to the pith. Column three (\code{d2pith}) gives the estimated distance to the pith (mm). Column four (\code{diam}) gives the diameter at breast height (DBH) in cm.
}
\source{
  Contributed by Stefan Klesse
}
\references{
  Klesse, S., Babst, F., Lienert, S., Spahni, R., Joos, F., Bouriaud, O., Carrer, M., Filippo, A.D., Poulter, B., Trotsiuk, V., Wilson, R., Frank, D.C. (2018) A combined tree ring and vegetation model assessment of European forest growth sensitivity to interannual climate variability. \emph{Global Biogeochemical Cycles} 32, 1226 -- 1240.
}
\keyword{datasets}
