\encoding{UTF-8}
\name{zof.rwl}
\docType{data}
\alias{zof.rwl}
\title{ European Beech Ring Widths from Zofingen, Switzerland }
\description{
  This data set includes ring-width measurements for European Beech (\emph{Fagus sylvatica}) increment cores collected at the Zofingen in Switzerland.  There are 61 series.
}
\usage{data(zof.rwl)}
\format{
  A \code{data.frame} containing 61 ring-width series in columns and 156
  years in rows.
}
\source{
  Contributed by Stefan Klesse
}
\references{
  Klesse, S., Babst, F., Lienert, S., Spahni, R., Joos, F., Bouriaud, O., Carrer, M., Filippo, A.D., Poulter, B., Trotsiuk, V., Wilson, R., Frank, D.C. (2018) A combined tree ring and vegetation model assessment of European forest growth sensitivity to interannual climate variability. \emph{Global Biogeochemical Cycles} 32, 1226 -- 1240.
}
\keyword{datasets}
